Move response-actions, response-navigation, response-view under inspector
