+++ projects/packages/forms/src/dashboard/components/inspector/response-actions/index.tsx
@@ -15,12 +15,12 @@ import {
 	deleteAction,
 	markAsReadAction,
 	markAsUnreadAction,
-} from '../../inbox/stage/actions.tsx';
+} from '../../../inbox/stage/actions.tsx';
 /**
  * Types
  */
-import type { FormResponse } from '../../../types/index.ts';
-import type { Registry } from '../../inbox/stage/types.tsx';
+import type { FormResponse } from '../../../../types/index.ts';
+import type { Registry } from '../../../inbox/stage/types.tsx';
 
 type ResponseNavigationProps = {
 	onActionComplete?: ( response: FormResponse ) => void;
