+++ projects/packages/forms/src/dashboard/components/inspector/response-view/field-email/index.tsx
@@ -1,7 +1,7 @@
 import {
 	__experimentalHStack as HStack, // eslint-disable-line @wordpress/no-unsafe-wp-apis
 } from '@wordpress/components';
-import CopyClipboardButton from '../../../components/copy-clipboard-button';
+import CopyClipboardButton from '../../../copy-clipboard-button';
 import './style.scss';
 
 const FieldEmail = ( { email } ) => {
