+++ projects/packages/forms/src/dashboard/components/inspector/response-view/field-image-select/index.tsx
