+++ projects/packages/forms/src/dashboard/components/inspector/response-view/field-preview/style.scss
