+++ projects/packages/forms/src/dashboard/components/inspector/single.tsx
@@ -5,9 +5,9 @@ import {
 import { useCallback, useEffect, useRef, useState } from '@wordpress/element';
 import useResponseNavigation from '../../hooks/use-response-navigation.ts';
 import Page from '../page/index.tsx';
-import ResponseActions from '../response-actions/index.tsx';
-import ResponseNavigation from '../response-navigation/index.tsx';
 import { ResponseViewBody } from './index.tsx';
+import ResponseActions from './response-actions/index.tsx';
+import ResponseNavigation from './response-navigation/index.tsx';
 
 /**
  * Single response component for dataviews.
